fix: use const over let
and updated project
