+++ .eslintignore
@@ -1,4 +1,6 @@
 node_modules
 dist*
+cjs
+esm
 lib*
 out*
