+++ .eslintrc.json
@@ -2,7 +2,20 @@
   "extends": "plugin:harmony/latest",
   "root": true,
   "env": {
+    "es6": true,
     "node": true,
-    "es6": true
-  }
+    "jest": true
+  },
+  "overrides": [
+    {
+      "extends": "plugin:harmony/ts-recommended",
+      "files": [
+        "*.ts",
+        "*.tsx"
+      ],
+      "rules": {
+        "no-use-before-define": "off"
+      }
+    }
+  ]
 }
