+++ .releaserc.json
@@ -1,0 +1,7 @@
+{
+  "branches": [
+    "main",
+    "next",
+    "beta"
+  ]
+}
