--- .vscode/tasks.json
@@ -1,27 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
-  "version": "0.1.0",
-  "command": "npm",
-  "args": [
-    "run-script"
-  ],
-  "isShellCommand": true,
-  "showOutput": "always",
-  "suppressTaskName": true,
-  "tasks": [
-    {
-      "taskName": "watch",
-      "args": [
-        "watch"
-      ]
-    },
-    {
-      "taskName": "npm run build",
-      "isBuildCommand": true,
-      "args": [
-        "build"
-      ]
-    }
-  ]
-}
