--- jest.config.js
@@ -1,4 +1,0 @@
-const common = require('@unional/devpkg-node/simple/config/jest.common')
-module.exports = Object.assign(common, {
-  'testEnvironment': 'node'
-})
