+++ jest.config.json
@@ -1,0 +1,47 @@
+{
+  "collectCoverageFrom": [
+    "<rootDir>/src/**/*.[jt]s",
+    "!<rootDir>/src/bin.[jt]s",
+    "!<rootDir>/src/type-checker/*"
+  ],
+    "reporters": [
+      "default",
+      "jest-progress-tracker",
+      [
+        "jest-junit",
+        {
+          "outputDirectory": ".reports/junit",
+          "outputName": "js-test-results.xml"
+        }
+      ]
+    ],
+      "roots": [
+        "<rootDir>/ts"
+      ],
+        "testEnvironment": "node",
+          "testMatch": [
+            "**/?(*.)+(spec|test|integrate|accept|system|unit).[jt]s?(x)"
+          ],
+            "transform": {
+    "^.+\\.(t|j)sx?$": [
+      "@swc/jest"
+    ]
+  },
+  "watchPlugins": [
+    "jest-watch-suspend",
+    "jest-watch-typeahead/filename",
+    "jest-watch-typeahead/testname",
+    [
+      "jest-watch-toggle-config",
+      {
+        "setting": "verbose"
+      }
+    ],
+    [
+      "jest-watch-toggle-config",
+      {
+        "setting": "collectCoverage"
+      }
+    ]
+  ]
+}
