+++ package.json
@@ -15,34 +15,52 @@
     "name": "Homa Wong (unional)",
     "email": "homawong@gmail.com"
   },
+  "main": "cjs/index.js",
+  "module": "esm/index.js",
+  "jsnext:main": "esm/index.js",
+  "typings": "cjs/index.d.ts",
   "files": [
-    "dist",
-    "lib",
-    "libm"
+    "cjs",
+    "dist"
+    "esm",
+    "ts"
   ],
-  "main": "lib/index.js",
-  "module": "libm/index.js",
-  "jsnext:main": "libm/index.js",
-  "typings": "lib/index.d.ts",
   "scripts": {
     "build": "tsc -p tsconfig.es5.json && tsc -p tsconfig.esnext.json",
     "bundle": "webpack",
-    "clean": "rimraf lib",
-    "dc": "npm run dependency-check",
+    "clean": "rimraf lib dist cjs esm",
     "dependency-check": "dependency-check . --unused --no-dev && dependency-check . --missing --no-dev",
-    "lint": "tslint -p tsconfig.json",
+    "lint": "eslint --ext=js,ts .",
     "prepublishOnly": "npm run build && npm run bundle",
     "test": "jest",
     "verify": "npm run lint && npm run build && npm run dependency-check && npm run bundle && jest --coverage",
     "watch": "jest --watch"
   },
   "devDependencies": {
-    "@unional/devpkg-node": "^1.3.0",
+    "@commitlint/cli": "^16.0.0",
+    "@commitlint/config-conventional": "^16.0.0",
+    "@swc/core": "^1.2.125",
+    "@swc/jest": "^0.2.15",
+    "@types/jest": "^27.4.0",
     "assertron": "^7.0.2",
     "colormap": "^2.3.1",
+    "dependency-check": "^4.1.0",
+    "eslint": "^8.6.0",
+    "eslint-plugin-harmony": "^6.0.0",
+    "husky": "^7.0.4",
+    "jest": "^27.4.5",
+    "jest-junit": "^13.0.0",
+    "jest-progress-tracker": "^3.0.1",
+    "jest-validate": "^27.4.2",
+    "jest-watch-repeat": "^2.0.0",
+    "jest-watch-suspend": "^1.1.2",
+    "jest-watch-toggle-config": "^2.0.1",
+    "jest-watch-typeahead": "^1.0.0",
     "param-case": "^2.1.1",
     "pascal-case": "^2.0.1",
+    "rimraf": "^3.0.2",
     "ts-loader": "^5.0.0",
+    "typescript": "^4.5.4",
     "webpack": "^4.31.0",
     "webpack-cli": "^3.3.2"
   }
