+++ ts/createColorsFromMap.spec.ts
