+++ ts/createColorsFromMap.ts
