+++ tsconfig.base.json
@@ -14,6 +14,6 @@
     "strict": true
   },
   "files": [
-    "src/index.ts"
+    "ts/index.ts"
   ]
 }
