+++ tsconfig.json
@@ -1,15 +1,7 @@
 {
   "extends": "./tsconfig.esnext.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "plugins": [
-      {
-        "name": "tslint-language-service"
-      }
-    ]
-  },
   "include": [
-    "src",
+    "ts",
     "typings"
   ]
 }
